refactor(bpe): move code next to its users
